@@ -1,41 +1,30 @@
-[tool.poetry]
+[project]
 name = "libeq"
-version = "0.1.0"
-description = ""
-authors = ["Lorenzo Castellino <castellino96@gmail.com>"]
+version = "0.3.0"
+description = "Library for solving equilibrium in solution problems"
+authors = [{name="Lorenzo Castellino", email="<castellino96@gmail.com>"},
+           {name="Salvador Blasco", email="<salvador.blasco@uv.es>"}]
 readme = "README.md"
 packages = [{ include = "libeq", from = "src" }]
+requires-python = ">= 3.10"
+dependencies = [
+ "numpy",
+]
+
+[project.scripts]
+libeq = "..src"
 
 [tool.pytest.ini_options]
 pythonpath = ["src"]
 
-[tool.poetry.dependencies]
-python = ">=3.9,<3.13"
-numpy = "^1.26.4"
-pydantic = "^2.6.4"
-pydantic-numpy = "^4.2.0"
+[tool.run.ini_options]
+pythonpath = ["src"]
 
-[tool.poetry.group.dev.dependencies]
-matplotlib = "^3.8.3"
-pandas = "^2.2.1"
-pytest = "^8.1.1"
-ipykernel = "^6.29.3"
-pyinstrument = "^4.6.2"
-ruff = "^0.3.4"
-inline-snapshot = "^0.8.2"
+[tool.mypy]
+mypy_path = "src"
 
-[tool.poetry.group.docs.dependencies]
-mkdocs-material = "^9.5.15"
-mkdocstrings = { extras = ["python"], version = "^0.24.1" }
-mkdocs-gen-files = "^0.5.0"
-mkdocs-literate-nav = "^0.6.1"
-mkdocs-section-index = "^0.3.8"
-markdown-katex = "^202112.1034"
-mkdocs-glightbox = "^0.3.7"
-black = "^24.3.0"
-griffe-fieldz = "^0.1.2"
-
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
+[dependency-groups]
+dev = [
+    "numpy>=2.2.6",
+    "pytest>=8.4.2",
+]
