@@ -1,5 +1,17 @@
+"""
+General functions for nonlinear fitting.
+"""
+
+
+from typing import Tuple, Dict, List, Final
+
 import numpy as np
+from numpy.typing import NDArray
+
 from libeq.excepts import TooManyIterations
+
+
+FArray = NDArray[float]
 
 
 def levenberg_marquardt(x0, y, f, free_conc, jacobian, weights, capping=None, **kwargs):
@@ -61,7 +73,7 @@ def levenberg_marquardt(x0, y, f, free_conc, jacobian, weights, capping=None, **
     n_points = len(y)
     n_vars = len(x0)
     sigma_hist = []
-
+    # breakpoint()
     # import pudb
     # pudb.set_trace()
 
