@@ -1,1 +1,2 @@
-from .bstac import parse_BSTAC_file  # noqa: F401
+from .bstac import parse_BSTAC_file
+from .superquad import parse_superquad_file
